requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.24.0.post1
redis[hiredis]==5.0.1
hvac==1.2.1
pytest==7.4.3
plexapi==4.15.4
python-json-logger==2.0.7
aiohttp[speedups]==3.9.1
asyncio-throttle==1.0.2
pydantic[email]==2.5.2
prometheus-client==0.19.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx[http2]==0.25.2
discord.py==2.3.2
aioredis==2.0.1
numpy==1.26.2
python-jose[cryptography]==3.3.0
tenacity==8.2.3
bcrypt==4.1.2
cryptography==41.0.7
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
sentry-sdk==1.35.0
statsd==4.0.1
opentelemetry-instrumentation-aiohttp==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-otlp==1.21.0
pytest-asyncio==0.23.2
pytest-cov==4.1.0
asyncpg==0.29.0
orjson==3.9.10
# Added performance and security packages
ujson==5.8.0
cchardet==2.1.7
aiodns==3.1.1
certifi==2023.11.17

# Development tools
black==23.11.0
isort==5.12.0
mypy==1.7.1
bandit==1.7.5
safety==2.3.5
semgrep==1.41.0
pytest-sugar==0.9.7
pytest-mock==3.12.0
pytest-xdist==3.3.1
types-redis==4.6.0.11
types-requests==2.31.0.10
